config_module/requirements.txt
pytest==8.3.2
PyYAML==6.0.1
configparser==7.0.0
